completed assignment3

+++ Fall 2021 CS 5007/assignments/Assignment3/ITExamGrade.py
@@ -5,7 +5,7 @@ from AnalyticalProgrammingExam import AnalyticalProgrammingExam
 
 def fillITExam(oneExamObject):
     #  Prompt user for title
-    title_flag = False #  True when title is valid
+    title_flag = False  # True when title is valid
 
     while not title_flag:
         title = input("\nEnter the exam title: ")
@@ -16,7 +16,7 @@ def fillITExam(oneExamObject):
 
     if isinstance(oneExamObject, MultipleChoiceExam):
         # Prompt user for total mcquestions
-        num_flag = False #  True when total number of mcq is valid
+        num_flag = False  # True when total number of mcq is valid
 
         while not num_flag:
             num_mcq = int(input("Total Number of Multiple Choice: "))
@@ -26,7 +26,7 @@ def fillITExam(oneExamObject):
                 print("\nSorry! The number is not valid. Please enter it again.")
 
         # Prompt user for num of correct mcquestions
-        num_correct_flag = False #  True when number of correct mcq is valid
+        num_correct_flag = False  # True when number of correct mcq is valid
 
         while not num_correct_flag:
             num_correct_mcq = int(input("Total Number of Correct Multiple Choice's Questions: "))
@@ -37,10 +37,10 @@ def fillITExam(oneExamObject):
 
     elif isinstance(oneExamObject, TechnicalWritingExam):
         #  Prompt user for grammer score
-        grammer_flag = False #  True when grammer score is valid
+        grammer_flag = False  # True when grammer score is valid
 
         while not grammer_flag:
-            gram_score = int(input("Score of Grammer Portion: "))
+            gram_score = float(input("Score of Grammer Portion: "))
             if oneExamObject.setGrammerScore(gram_score):
                 grammer_flag = True
             else:
@@ -50,17 +50,17 @@ def fillITExam(oneExamObject):
         sent_str_flag = False  # True when sentence structure score is valid
 
         while not sent_str_flag:
-            sent_str_score = int(input("Score of Sentence Structure Portion: "))
+            sent_str_score = float(input("Score of Sentence Structure Portion: "))
             if oneExamObject.setSentenceStructureScore(sent_str_score):
                 sent_str_flag = True
             else:
                 print("\nSorry! The score is not valid. Please enter it again.")
 
         #  Prompt user for content score
-        content_flag = False #  True when content score is valid
+        content_flag = False  # True when content score is valid
 
         while not content_flag:
-            content_score = int(input("Score of Content Portion: "))
+            content_score = float(input("Score of Content Portion: "))
             if oneExamObject.setContentScore(content_score):
                 content_flag = True
             else:
@@ -68,20 +68,20 @@ def fillITExam(oneExamObject):
 
     elif isinstance(oneExamObject, AnalyticalProgrammingExam):
         #  Prompt user for short ans score
-        short_ans_flag = False #  True when short answer score is valid
+        short_ans_flag = False  # True when short answer score is valid
 
         while not short_ans_flag:
-            short_ans_score = int(input("Score of Short Answer Section: "))
+            short_ans_score = float(input("Score of Short Answer Section: "))
             if oneExamObject.setShortAnswerScore(short_ans_score):
                 short_ans_flag = True
             else:
                 print("\nSorry! The score is not valid. Please enter it again.")
 
         #  Prompt user for programming score
-        prog_flag = False #  True when programming score is valid
+        prog_flag = False  # True when programming score is valid
 
         while not prog_flag:
-            prog_score = int(input("Score of Programming Section: "))
+            prog_score = float(input("Score of Programming Section: "))
             if oneExamObject.setProgrammingScore(prog_score):
                 prog_flag = True
             else:
